fix: airfleet helper
