+++ inc/Helpers/Helper.php
@@ -3,21 +3,24 @@
 namespace Airfleet\Framework\Helpers;
 
 abstract class Helper {
-	protected static $instance;
+	protected static array $instances = [];
 
-	abstract protected static function create_instance(): static;
+	abstract protected static function instance_name(): string;
+	abstract protected static function create_instance(): mixed;
 
-	public static function __callStatic( string $method, array $args ): mixed {
+	public static function __callStatic( string $method, array $args ) {
 		$instance = static::instance();
 
 		return call_user_func_array( [ $instance, $method ], $args );
 	}
 
-	protected static function instance(): static {
-		if ( ! isset( static::$instance ) ) {
-			static::$instance = static::create_instance();
+	protected static function instance(): mixed {
+		$name = static::instance_name();
+
+		if ( ! isset( static::$instances[ $name ] ) ) {
+			static::$instances[ $name ] = static::create_instance();
 		}
 
-		return static::$instance;
+		return static::$instances[ $name ];
 	}
 }
